Refactor bank offset function to normalizations.py
