+++ total_scattering/reduction/normalizations.py
@@ -1,9 +1,12 @@
 # 3rd-party imports
 from mantid.api import mtd
 from mantid.dataobjects import Workspace2D
-from mantid.simpleapi import Divide, RenameWorkspace
+from mantid.simpleapi import (
+    CloneWorkspace, CreateWorkspace, DeleteWorkspace, Divide, MatchSpectra,
+    RenameWorkspace)
 
 # standard imports
+import numpy as np
 from typing import Union
 
 
@@ -48,18 +51,66 @@ def to_absolute_scale(
                     OutputWorkspace=output_workspace)
 
 
-# TODO #354 (This mock returns 1.0 everywhere)
 def calculate_fitted_levels(
         input_workspace: Union[str, Workspace2D],
         q_ranges: dict,
         output_workspace: str = 'fitted_levels') -> Workspace2D:
-    r"""
-    :param input_workspace:
-    :param output_workspace:
+    r"""Fits a horizontal line to each bank and region specified in q_ranges.
+    Scales the full bank data in input_workspace by the offset (fitted level)
+    and returns a new workspace with the scaled data (s_q_norm)
+
+    :param input_workspace: workspace containing S(Q) with spectra for each
+    bank, in MomentumTransfer units
+    :param q_ranges: dictionary of bank number with tuple range of Q fitting
+    :param output_workspace: name of output workspace
+    :return: Clone of input_workspace scaled by fitted level for banks
+    specified in q_ranges
     """
-    return Divide(LHSWorkspace=input_workspace,
-                  RHSWorkspace=input_workspace,
-                  OutputWorkspace=output_workspace)
+    input_workspace = mtd[str(input_workspace)]
+
+    # Check units (and for now, throw error if not in MomentumTransfer)
+    if input_workspace.getAxis(0).getUnit().unitID() != "MomentumTransfer":
+        raise RuntimeError(
+            "S(Q) workspace was expected to be in MomentumTransfer, "
+            "but has units '{}'".format(
+                input_workspace.getAxis(0).getUnit().unitID()))
+
+    # Setup the output workspace. This gets returned as-is if q_ranges is empty
+    s_q_norm = CloneWorkspace(InputWorkspace=input_workspace,
+                              OutputWorkspace=str(output_workspace))
+
+    # Perform the horizontal fitting to each bank in q_ranges
+    for key, value in q_ranges.items():
+        # get corresponding workspace index number in S(Q) from the bank number
+        ws_index = key - 1
+
+        start_index = input_workspace.yIndexOfX(value[0], ws_index) + 1
+        end_index = input_workspace.yIndexOfX(value[1], ws_index) + 1
+        # Extract the bank between the fit regions
+        bank_x = input_workspace.readX(ws_index)[start_index:end_index]
+        bank_y = input_workspace.readY(ws_index)[start_index:end_index]
+        bank_e = input_workspace.readE(ws_index)[start_index:end_index]
+
+        # Create reference spectrum which is flat line at 1.0
+        ref_y = np.zeros(bank_x.shape) + 1.0
+
+        CreateWorkspace(np.concatenate((bank_x, bank_x)),
+                        np.concatenate((ref_y, bank_y)),
+                        np.concatenate((ref_y, bank_e)),
+                        NSpec=2,
+                        OutputWorkspace="__tmp_bank_fit")
+        match_result = MatchSpectra(InputWorkspace="__tmp_bank_fit",
+                                    OutputWorkspace="__tmp_bank_fit",
+                                    ReferenceSpectrum=1,
+                                    CalculateOffset=True,
+                                    CalculateScale=False)
+        offset = match_result.Offset[1]
+        # scale full bank data by offset
+        s_q_norm.setY(ws_index, s_q_norm.dataY(ws_index) * (1.0 / offset))
+
+        DeleteWorkspace("__tmp_bank_fit")
+
+    return s_q_norm
 
 
 def to_f_of_q(
